Changed failed writers handling.
